2023 day 5 part 1

+++ src/main/scala/2023/Day_5.scala
@@ -5,8 +5,8 @@ import util.FileReader
 import scala.io.Source
 
 object Day_5:
-  case class Input(seeds: List[Int], maps: List[List[PlantingMap]])
-  case class PlantingMap(destinationRange: Int, sourceRange: Int, rangeLength: Int)
+  case class Input(seeds: List[BigInt], maps: List[List[PlantingMap]])
+  case class PlantingMap(destinationRange: BigInt, sourceRange: BigInt, rangeLength: BigInt)
 
   val index = 5
 
@@ -17,9 +17,9 @@ object Day_5:
         .filter(c => c == ' ' || c == '\n' || c.isDigit)
         .split("\n\n")
 
-    val seeds = separated.head.split(' ').filter(_.nonEmpty).map(_.toInt)
+    val seeds = separated.head.split(' ').filter(_.nonEmpty).map(BigInt(_))
     val maps = separated.tail
-      .map(_.split("\n").filter(_.nonEmpty).map(_.split(' ').map(_.toInt)))
+      .map(_.split("\n").filter(_.nonEmpty).map(_.split(' ').map(BigInt(_))))
       .map(_.filter(_.nonEmpty).toList.map {
         case Array(a, b, c) => PlantingMap(a, b, c)
         case _              => throw new Error("could not match")
@@ -28,7 +28,7 @@ object Day_5:
 
     Input(seeds.toList, maps)
 
-  def part1(input: Input): Int =
+  def part1(input: Input): BigInt =
     input.maps.foldLeft(input.seeds)((acc, maps) => {
       val res = acc.map(seed =>
         maps.find(map => seed >= map.sourceRange && seed < map.sourceRange + map.rangeLength) match
